Patrolling Generator in separate file and adding PatrollingAlgorithm base class
