+++ Assets/Scripts/Algorithms/IPatrollingAlgorithm.cs
@@ -2,7 +2,8 @@ using Maes.Map;
 
 namespace Maes.Algorithms
 {
-    public interface IPatrollingAlgorithm : IAlgorithm {
+    public interface IPatrollingAlgorithm : IAlgorithm
+    {
         void SetPatrollingMap(PatrollingMap map);
     }
 }
