+++ Assets/Scripts/Map/RobotSpawners/PatrollingRobotSpawner.cs
@@ -6,19 +6,19 @@ using Maes.Map;
 using Maes.Map.MapGen;
 using Maes.Robot;
 
+using UnityEngine;
+
 namespace MAES.Map.RobotSpawners
 {
     public class PatrollingRobotSpawner : RobotSpawner<IPatrollingAlgorithm>
     {
-        [CanBeNull] private PatrollingMap _patrollingMap;
+        public PatrollingMap PatrollingMap;
         protected override MonaRobot CreateRobot(float x, float y, float relativeSize, int robotId, IPatrollingAlgorithm algorithm,
             SimulationMap<Tile> collisionMap, int seed)
         {
             var robot = base.CreateRobot(x, y, relativeSize, robotId, algorithm, collisionMap, seed);
 
-            _patrollingMap ??= new PatrollingMap(collisionMap);
-
-            algorithm.SetPatrollingMap((PatrollingMap)_patrollingMap.Clone());
+            algorithm.SetPatrollingMap((PatrollingMap)PatrollingMap.Clone());
             
             return robot;
         }
