+++ Assets/Scripts/PatrollingAlgorithm/ConscientiousReactiveAlgorithm.cs
@@ -11,18 +11,18 @@ using UnityEngine;
 
 namespace Maes.PatrollingAlgorithm.ConscientiousReactive
 {
-    public class ConscientiousReactiveAlgorithm : IPatrollingAlgorithm
+    public class ConscientiousReactiveAlgorithm : MAES.PatrollingAlgorithm.PatrollingAlgorithm
     {
         private Robot2DController _controller;
         private IReadOnlyList<Vertex> _vertices;
         private Vertex _currentVertex;
         private bool _isPatrolling = false;
-        public void SetPatrollingMap(PatrollingMap map)
+        public override void SetPatrollingMap(PatrollingMap map)
         {
             _vertices = map.Verticies;
         }
 
-        public string GetDebugInfo()
+        public override string GetDebugInfo()
         {
             return 
             "Conscientious Reactive Algorithm\n" + 
@@ -30,12 +30,12 @@ namespace Maes.PatrollingAlgorithm.ConscientiousReactive
             $"Init done:  {_isPatrolling}\n";
         }
 
-        public void SetController(Robot2DController controller)
+        public override void SetController(Robot2DController controller)
         {
             _controller = controller;
         }
 
-        public void UpdateLogic()
+        public override void UpdateLogic()
         {
             if(!_isPatrolling){
                 var vertex = GetClosestVertex();
