+++ Assets/Scripts/Simulation/ExplorationSimulation.cs
@@ -11,7 +11,7 @@ using UnityEngine;
 
 namespace Maes
 {
-    public sealed class ExplorationSimulation : SimulationBase<ExplorationSimulation, ExplorationVisualizer, ExplorationCell, ExplorationTracker, ExplorationInfoUIController, IExplorationAlgorithm, ExplorationSimulationScenario>
+    public sealed class ExplorationSimulation : SimulationBase<ExplorationSimulation, ExplorationVisualizer, ExplorationCell, ExplorationTracker, ExplorationInfoUIController, IExplorationAlgorithm, ExplorationSimulationScenario, ExplorationRobotSpawner>
     {
         public ExplorationTracker ExplorationTracker { get; set; }
