+++ Assets/Scripts/Simulation/PatrollingSimulation.cs
@@ -1,6 +1,13 @@
+using System.Collections.Generic;
+
 using Maes.Algorithms;
+using Maes.Map;
 using Maes.Map.MapGen;
+using Maes.Map.MapPatrollingGen;
+
 using MAES.Map.RobotSpawners;
+
+using Maes.Robot;
 using Maes.Simulation;
 using MAES.Simulation.SimulationScenarios;
 using Maes.Statistics;
@@ -13,7 +20,7 @@ using UnityEngine;
 
 namespace Maes
 {
-    public sealed class PatrollingSimulation : SimulationBase<PatrollingSimulation, PatrollingVisualizer, Tile, PatrollingTracker, PatrollingInfoUIController, IPatrollingAlgorithm, PatrollingSimulationScenario>
+    public sealed class PatrollingSimulation : SimulationBase<PatrollingSimulation, PatrollingVisualizer, Tile, PatrollingTracker, PatrollingInfoUIController, IPatrollingAlgorithm, PatrollingSimulationScenario, PatrollingRobotSpawner>
     {
         public GameObject patrollingVisualizerPrefab;
         public PatrollingVisualizer patrollingVisualizer;
@@ -23,6 +30,8 @@ namespace Maes
         public override PatrollingVisualizer Visualizer => patrollingVisualizer;
 
         public override PatrollingTracker Tracker => PatrollingTracker;
+        
+        private PatrollingMap _patrollingMap;
         
         protected override void AfterStart()
         {
@@ -35,9 +44,17 @@ namespace Maes
         public override void SetScenario(PatrollingSimulationScenario scenario)
         {
             base.SetScenario(scenario);
-
+            
             PatrollingTracker = new PatrollingTracker(scenario.RobotConstraints);
             patrollingVisualizer.SetMap(_collisionMap, _collisionMap.ScaledOffset);
+            patrollingVisualizer.SetPatrollingMap(_patrollingMap);
+        }
+
+        protected override void SpawnRobots(PatrollingSimulationScenario scenario)
+        {
+            _patrollingMap = scenario.PatrollingMapFactory(new PatrollingMapSpawner(), _collisionMap);
+            RobotSpawner.PatrollingMap = _patrollingMap;
+            base.SpawnRobots(scenario);
         }
 
         public override void OnDestory()
