+++ Assets/Scripts/Simulation/SimulationScenarios/PatrollingSimulationScenario.cs
@@ -1,5 +1,8 @@
 using Maes;
 using Maes.Algorithms;
+using Maes.Map;
+using Maes.Map.MapGen;
+using Maes.Map.MapPatrollingGen;
 using Maes.PatrollingAlgorithm.ConscientiousReactive;
 using Maes.Robot;
 
@@ -7,16 +10,21 @@ using UnityEngine;
 
 namespace MAES.Simulation.SimulationScenarios
 {
+    public delegate PatrollingMap PatrollingMapFactory(PatrollingMapSpawner generator, SimulationMap<Tile> map);
+    
     public sealed class PatrollingSimulationScenario : SimulationScenario<PatrollingSimulation, IPatrollingAlgorithm>
     {
+        
+        public PatrollingMapFactory PatrollingMapFactory { get; }
         public PatrollingSimulationScenario(
             int seed,
-            SimulationEndCriteriaDelegate<PatrollingSimulation>
-                hasFinishedSim=null, 
+            SimulationEndCriteriaDelegate<PatrollingSimulation> hasFinishedSim=null,
             MapFactory mapSpawner=null,
             RobotFactory<IPatrollingAlgorithm> robotSpawner=null,
             RobotConstraints? robotConstraints=null,
-            string statisticsFileName=null) 
+            string statisticsFileName=null,
+            PatrollingMapFactory patrollingMapFactory=null
+            ) 
             : base(seed,
                 robotSpawner ?? ((map, spawner) => spawner.SpawnRobotsTogether(map, seed, 1, Vector2Int.zero, (robotSeed) => new ConscientiousReactiveAlgorithm())),
                 hasFinishedSim, 
@@ -24,6 +32,7 @@ namespace MAES.Simulation.SimulationScenarios
                 robotConstraints,
                 statisticsFileName)
         {
+            PatrollingMapFactory = patrollingMapFactory ?? ((generator, map) => generator.GeneratePatrollingMapRetanglesBased(map));
         }
     }
 }
